requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0
pyyaml>=6.0
python-dateutil>=2.8
